chore: Renames command line args file.
